backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0  # For testing with in-memory SQLite

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0.0  # Lock to 4.x to avoid 5.x breaking changes
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# AI/ML
openai==1.3.5
google-generativeai>=0.8.0
scikit-learn==1.3.2

# Stock/Crypto Data
yfinance>=0.2.66
alpha-vantage==3.0.0

# Scheduling
apscheduler==3.10.4
python-dateutil==2.8.2
pytz==2023.3

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
py-vapid==1.9.0
pywebpush==1.14.0
slowapi==0.1.9
email-validator==2.1.0
pandas-market-calendars==4.4.1
finnhub-python==2.4.19
redis==5.0.1
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.40.0

# Testing (Phase 6)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Scalability (Phase 7)
celery==5.3.4
flower==2.0.1

# Business Model (Phase 8)
stripe==7.7.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
